Renames `shared/index` to `shared/MainApp`
Also uses the ‘App’ alias to resolve the correct version of the app
