+++ client/index.js
@@ -1,6 +1,3 @@
-/* eslint-disable global-require */
-/* eslint-disable no-console */
-
 import React from 'react';
 import { render } from 'react-dom';
 import BrowserRouter from 'react-router-dom/BrowserRouter';
@@ -12,10 +9,10 @@ import { toJS } from 'mobx';
 import stringify from 'json-stringify-safe';
 import ReactGA from 'react-ga';
 import Store from 'store';
-import config from 'utils/config';
+import App from 'App';
 
+import config from 'utils/config';
 import ReactHotLoader from './components/ReactHotLoader';
-import App from '../shared';
 
 // Initialize Google Analytics
 const gaId = config('gaId');
@@ -91,7 +88,7 @@ if (process.env.BUILD_FLAG_IS_DEV === 'true' && module.hot) {
   // Accept changes to this file for hot reloading.
   module.hot.accept('./index.js');
   // Any changes to our App will cause a hotload re-render.
-  module.hot.accept(() => renderApp(require('../shared').default));
+  module.hot.accept(() => renderApp(require('App').default));
 
   const consoleWarn = console.warn;
